refactor: remove AutoTracker class and associated types for cleaner codebase
